+++ modules/tie/client.go
@@ -5,10 +5,10 @@ import (
 	"errors"
 	"fmt"
 
-	"git.sr.ht/~uid/tie/client"
-	"git.sr.ht/~uid/tie/tiedb"
 	"github.com/deepnoodle-ai/risor/v2/pkg/object"
 	"github.com/deepnoodle-ai/risor/v2/pkg/op"
+	"github.com/uidbz/tie/client"
+	"github.com/uidbz/tie/tiedb"
 )
 
 const ClientType object.Type = "tie.client"